+++ src/main/java/com/poc/repository/ApsrtcRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.poc.model.Ticket;
 
-public interface ApsrtcRepository extends JpaRepository<Ticket, Long>{
+public interface ApsrtcRepository extends JpaRepository<Ticket, Integer>{
 
 }
